Improved driver implementation.
Driver now uses a lazily initialized static terminfo database, which is
validated on calling Driver::new() to guarantee that required device functions
are available. In addition this removes any need to clone the large underlying
hashmap, which would be an expensive operation.

Minor renaming of internal functions/methdods.

Changes and additions to comments.