First draft of a new "ui" module
This module contains a convenience layer to display various widgets,
such as windows, buttons etc. It also manages these elements' alignment,
size and all.

This is a squash commit of a WIP branch I have been maintaining for a
couple of weeks to support the development of another projet. Until now,
I had no idea where I was going with the UI module, but now things are
starting to take shape and I'm gearing up for a first PR.

In addition to the new UI modules, a couple of changes have been made to
core:

`Coordinate` removal: The purpose of `Coordinate` was exactly the same
as `Option`. Replaced usage of `Coordinate` with `Option`.

`Cursor` un-traiting: The `Cursor` struct had an overly complex trait
structure that it was te only one to use. I moved in all directly in
`Cursor`.

Positioning and size consolidation: With the UI module, position and
size take a new, heightened importance. I renamed `Pair` type to `Pos`,
added a new `Size` type and added `HasSize` and `HasPosition` traits.
Only `HasSize` is used in core (`HasPosition` is used in `ui`), but it
felt weird to have both these traits at difference places, so all in
core...

New `CellAccessor` trait: With the new ui module, we need a common
interface to "write cells to an arbitrary surface", with that surface
being either a sub-window or the terminal itself.

On the "ui" side, we have (for now):

widget: a surface with a size and position in an arbitrary parent. We
draw in it like we would in a terminal, and then draw the widget itself
in another `CellAccessor`.

painter: A convenience trait that applies to any `CellAccessor` with
useful methods for easily painting multiple cells, such a `printline()`
or `draw_box()`.

button: a widget that draws its text wrapped around `<>` brackets and
highlights a chosen accel char.

layout: Helpers to easily align widgets inside each other.

All of this is heavily WIP, but I think it's advanced enough to open a
PR to start a discussion about the API design.