Flatten Cell's structure
Remove `Style` in order to flatten `Cell` structure. Rather than `fg`
and `bg` each having both a `Color` and a `Attr`, they both become
straight up `Color` and we add a new `attrs` attribute to the `Cell`.

This is more in line with the underlying driver (the `term` module)
which doesn't apply its attribute to foreground or background. It's
always the whole cell that it affected.

It also makes working with the API a bit lighter. It could become
cumbersome to have to involve both `Style` and `Color` in order to
change foreground or background colors.