Major internal and public API changes.
Removed the ui module for the moment, it will come back soon but with breaking
changes happening all over the place I can't keep it up to date as I get the
core module on track to becoming stable.

Several name changes so far, and I have removed many of the `with_cell` and
`with_char` functions as I believe they are only adding more code to the
library with little added usefulness.

Cursor position optimization has been removed for the moment as well, as I want
to focus on ensuring a solid public api at the moment before getting bogged
down with various optimizations.

More breaking changes are to come.