Terminal::get_event to uses Driver::get_event
This refactors Terminal::get_event to use the driver to get an event
from a string received from the terminal. It moves most of read_events'
functionality into get_event and removes read_event. The per-terminal
event buffer is removed, as get_event will return an event as soon as it
happens, so there's no need to buffer events.